split tests and cleanup imports
